+++ component/pl_mutex.cpp
@@ -1,5 +1,6 @@
 #include "pl_mutex.h"
 #include "esp_check.h"
+#include <cstdlib>
 
 //==============================================================================
 
@@ -11,19 +12,22 @@ namespace PL {
 
 //==============================================================================
 
-Mutex::Mutex() : mutex(xSemaphoreCreateRecursiveMutex()) {}
+Mutex::Mutex() : mutex(xSemaphoreCreateRecursiveMutex()) {
+  if (!mutex) {
+    ESP_LOGE(TAG, "mutex creation failed");
+    abort();
+  }
+}
 
 //==============================================================================
 
 Mutex::~Mutex() {
-  if (mutex)
-    vSemaphoreDelete(mutex);
+  vSemaphoreDelete(mutex);
 }
 
 //==============================================================================
 
 esp_err_t Mutex::Lock(TickType_t timeout) {
-  ESP_RETURN_ON_FALSE(mutex, ESP_ERR_INVALID_STATE, TAG, "mutex is null");
   ESP_RETURN_ON_FALSE(!xPortInIsrContext(), ESP_ERR_INVALID_STATE, TAG, "calling mutex lock from ISR");
   if (xSemaphoreTakeRecursive(mutex, timeout))
     return ESP_OK;
@@ -35,7 +39,6 @@ esp_err_t Mutex::Lock(TickType_t timeout) {
 //==============================================================================
 
 esp_err_t Mutex::Unlock() {
-  ESP_RETURN_ON_FALSE(mutex, ESP_ERR_INVALID_STATE, TAG, "mutex is null");
   ESP_RETURN_ON_FALSE(!xPortInIsrContext(), ESP_ERR_INVALID_STATE, TAG, "calling mutex unlock from ISR");
   ESP_RETURN_ON_FALSE(xSemaphoreGiveRecursive(mutex), ESP_FAIL, TAG, "semaphore give failed");
   return ESP_OK;